chore(cosmos): extract `app/upgrade.go`
